Clean up defines in makefile
Remove "MAKEFILE" from the CPP macro names in the makefile, and remove
BOARD and MAKEFILE_BOARD (the variable in makefile and the CPP macro,
respectively).  The removal of MAKEFILE_BOARD requires removing some
conditional compilation #ifdef's, but this firmware only supports the
Teensy 2.0, so there is no alternative code when BOARD is not
"teensy-2-0".