Removed references.md
This is less a references file and more a list of resources the original
found useful at some point during the project.  The few I thought were
specific and detailed enough to keep were (not all but mostly) dead
links, so I am removing the list instead of just pruning it.